Init hubs
