+++ src/Base/Service/Service.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\Database\Database.csproj" />
       <ProjectReference Include="..\Mq\Mq.csproj" />
       <ProjectReference Include="..\Types\Types.csproj" />
     </ItemGroup>
