+++ src/Base/Types/Exceptions/TypeException.cs
@@ -1,0 +1,12 @@
+namespace Types.Exceptions;
+
+public class TypeException: Exception
+{
+    public TypeException()
+    {
+    }
+    
+    public TypeException(string message) : base(message)
+    {
+    }
+}
