+++ src/Base/Types/Interfaces/Model/IUsername.cs
@@ -1,5 +1,4 @@
 using System.Text.RegularExpressions;
-using Types.Extensions;
 
 namespace Types.Interfaces.Model;
