+++ src/Services/Hubs/Hubs.API/Hubs.API.csproj
@@ -8,7 +8,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\Base\Database\Database.csproj"/>
         <ProjectReference Include="..\..\..\Base\Service\Service.csproj"/>
         <ProjectReference Include="..\Hubs.Domain\Hubs.Domain.csproj"/>
     </ItemGroup>
