--- src/Services/Hubs/Hubs.Domain/DataTransferObjects/LoginDto.cs
@@ -1,14 +1,0 @@
-using Types.Interfaces.Model;
-
-namespace Hubs.Domain.DataTransferObjects;
-
-public class LoginDto : IPassword
-{
-    /// <summary>
-    /// Identifier of user, either email or username
-    /// </summary>
-    public string Identifier { get; set; } = string.Empty;
-
-    /// <inheritdoc />
-    public string Password { get; set; } = string.Empty;
-}
