--- src/Services/Hubs/Hubs.Domain/DataTransferObjects/LoginResultDto.cs
@@ -1,7 +1,0 @@
-namespace Hubs.Domain.DataTransferObjects;
-
-public class LoginResultDto
-{
-    public required string Token { get; set; }
-    public required FullUserDto User { get; set; }
-}
