--- src/Services/Hubs/Hubs.Domain/DataTransferObjects/RegisterDto.cs
@@ -1,15 +1,0 @@
-using Types.Interfaces.Model;
-
-namespace Hubs.Domain.DataTransferObjects;
-
-public class RegisterDto : IUsername, IEmail, IPassword
-{
-    /// <inheritdoc />
-    public string Username { get; set; } = string.Empty;
-
-    /// <inheritdoc />
-    public string? Email { get; set; } = string.Empty;
-
-    /// <inheritdoc />
-    public string Password { get; set; } = string.Empty;
-}
