--- src/Services/Hubs/Hubs.Domain/Enums/UserState.cs
@@ -1,10 +1,0 @@
-namespace Hubs.Domain.Enums;
-
-public enum UserState
-{
-    Unknown,
-    NotActivated,
-    Activated,
-    Deleted,
-    Banned
-}
