+++ src/Services/Users/Users.Domain/DataTransferObjects/FullUserDto.cs
@@ -1,4 +1,3 @@
-using Types.Enums;
 using Types.Interfaces.Model;
 using Users.Domain.Enums;
