+++ src/Services/Users/Users.Domain/DataTransferObjects/SimpleUserDto.cs
